Update README.md
The Project on which I worked during my academic period is “Chronic Kidney Disease Prediction using Machine Learning. Machine Learning is used across many ranges around the world. The healthcare industry is no exclusion. Machine Learning can play an essential role in predicting presence/absence of locomotors disorders, CKD diseases and more. Such information, if predicted well in advance, can provide important intuitions to doctors who can then adapt their diagnosis and dealing per patient basis. Lifestyle diseases are common among the population today not only in India but also in almost every country. Lifestyle diseases are caused because of the habits that we have on a day to day basis. The way one lives his life is the major cause of it. It includes heart disease, hypertension, chronic kidney disease and liver disease, etc., which all may hear of.
CKD` disease has been a major life threatening problem in a human being. Due to several reasons such as eating habits, lack of exercise, becoming overweight, and smoking and unhealthy lifestyle habits cause’s ckd diseases. This motivates us to get a clear idea of the risk factors in our life regarding the cause of ckd disease. The proposed method suggests the levels of risk factors according the patients data, so that one can take care of their health properly in order to prevent the CKD disease. 

In this project work, proposed methodology is to segregate CKD disease patients based on their risk factors. Firstly applying the feature selection, this is to reduce the records of data and to train the model faster. The four classifiers are SVM, KNN, random forest and ADA Boost Models are applied. After applying the classifiers a proposed model will segregate the CKD disease patients based on their risk factors according to age of the patients. This will be helping for the doctors to analyze the risk factors of the patients. We work on predicting CKD Diseases in people using Machine Learning algorithms. Applied some the data analysis concept to analyse the data and perform the prediction whether patient has CKD or No CKD.